requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
anthropic>=0.25
openai>=1.20
google-generativeai>=0.5
python-dotenv>=1.0
pydantic>=2.0
aiohttp>=3.9
pytest-asyncio>=0.23
